Add files via upload
feat: Create homepage for Vacation Blog

Summarize changes in around 50 characters or less

This commit adds the initial structure and content for the homepage of the
Vacation Blog. The page includes a navigation bar with links to the home and
about pages, a header section introducing the blog, and a flexbox container
with information about two travel destinations - The Cromwell Las Vegas and
the Amalfi Coast.

Each destination has its own section with a brief description and an image. The
'Learn More' button in the header links to the contact page. Additionally, a
footer section is added with a 'Contact' button and social media icons.

This design aims to engage visitors with captivating visuals and provide easy
navigation to other sections of the blog.